Add third section and its styles
Created a div with a class .third-container, inside of it, another
wrapping div with a div and a button inside. the inner div contains an
h3 element and a p element. Both divs were set as flex-boxes. The outer
one with a padding, the inner, with some border radius and corresponding
colors set, also aligning (and justifying) its items.